+++ src/Core/Checkout/Customer/SalesChannel/ListAddressRoute.php
@@ -4,6 +4,7 @@ namespace Shopware\Core\Checkout\Customer\SalesChannel;
 
 use OpenApi\Annotations as OA;
 use Shopware\Core\Checkout\Customer\CustomerEntity;
+use Shopware\Core\Checkout\Customer\Event\AddressListingCriteriaEvent;
 use Shopware\Core\Framework\DataAbstractionLayer\EntityRepositoryInterface;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Filter\EqualsFilter;
@@ -13,7 +14,7 @@ use Shopware\Core\Framework\Routing\Annotation\LoginRequired;
 use Shopware\Core\Framework\Routing\Annotation\RouteScope;
 use Shopware\Core\Framework\Routing\Annotation\Since;
 use Shopware\Core\System\SalesChannel\SalesChannelContext;
-use Shopware\Storefront\Page\Address\Listing\AddressListingCriteriaEvent;
+use Shopware\Storefront\Page\Address\Listing\AddressListingCriteriaEvent as StorefrontAddressListingCriteriaEvent;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 
@@ -68,9 +69,8 @@ class ListAddressRoute extends AbstractListAddressRoute
             ->addAssociation('country')
             ->addFilter(new EqualsFilter('customer_address.customerId', $customer->getId()));
 
-        $this->eventDispatcher->dispatch(
-            new AddressListingCriteriaEvent($criteria, $context)
-        );
+        $this->eventDispatcher->dispatch(new StorefrontAddressListingCriteriaEvent($criteria, $context));
+        $this->eventDispatcher->dispatch(new AddressListingCriteriaEvent($criteria, $context));
 
         return new ListAddressRouteResponse($this->addressRepository->search($criteria, $context->getContext()));
     }
